GH-1494: Add helper
